QuestionsController

+++ Services/QuestionsServices.cs
@@ -35,12 +35,19 @@ namespace Project2.Services
 
         public Questions getQuestion(Guid questionId)
         {
-            return context.Questions.FirstOrDefault(x => x.QuestionId == questionId);
+            var response = context.Questions.FirstOrDefault(x => x.QuestionId == questionId);
+
+            return response;
         }
 
         public void updateQuestion(Guid questionId, QuestionsDto newQuestion)
         {
             var question = context.Questions.FirstOrDefault(x => x.QuestionId == questionId);
+
+            if (question == null)
+            {
+                return;
+            }
 
             question.ExamId = newQuestion.ExamId;
             question.Question = newQuestion.Question;
@@ -56,6 +63,12 @@ namespace Project2.Services
         public void deleteQuestion(Guid questionId)
         {
             var question = context.Questions.FirstOrDefault(x => x.QuestionId == questionId);
+
+            if (question == null)
+            {
+                return;
+            }
+
             context.Questions.Remove(question);
             context.SaveChanges();
         }
